Atoi fix

+++ iwstat.go
@@ -30,7 +30,7 @@ func Scan(r io.Reader) ([]IWStat, error) {
 			continue
 		}
 		var times [21]int
-		for i, idx := range []int{2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20} {
+		for i, idx := range []int64{2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20} {
 			v, err := strconv.Atoi(fields[idx])
 			if err != nil {
 				return nil, err
